keep track of pieces
