+++ src/types/puzzle-types.ts
@@ -22,3 +22,11 @@ export type PuzzleData = {
   fixedPieces: number[];
   removablePieces: number[];
 };
+
+// this is how we keep track of what pieces are on the board
+export type PieceState = {
+  isOnBoard: boolean;
+  variation: number | null; // variation index. unsure how to do this still
+  position: { row: number; col: number } | null;
+};
+export type PieceStatusMap = Record<number, PieceState>; // number = pieceId
